refactor: 패키지 구조 변경
- Person class 관련 클래스 생성
